Pientä refaktorointia ja koodin siivousta
