+++ src/app/components/table/getSortParts.tsx
@@ -1,0 +1,17 @@
+export const getSortParts = (sortStr: string, colId?: string) => {
+  const [orderBy, direction] = sortStr?.split(':') ?? '';
+
+  if (
+    (colId === undefined || colId === orderBy) &&
+    (direction === 'asc' || direction === 'desc')
+  ) {
+    return { orderBy, direction } as {
+      orderBy: string;
+      direction: 'asc' | 'desc';
+    };
+  }
+  return {
+    orderBy: undefined,
+    direction: undefined,
+  };
+};
