+++ src/app/components/table/list-table.tsx
@@ -14,6 +14,7 @@ import { TranslatedName, getTranslation } from '@/app/lib/common';
 import { Haku, getAlkamisKausi, Tila } from '@/app/lib/kouta-types';
 import { colors } from '@/app/theme';
 import { ExpandLess, ExpandMore, UnfoldMore } from '@mui/icons-material';
+import { getSortParts } from './getSortParts';
 
 type Column<P> = {
   title?: string;
@@ -131,24 +132,6 @@ const SortIcon = ({
     default:
       return <UnfoldMore />;
   }
-};
-
-export const getSortParts = (sortStr: string, colId?: string) => {
-  const [orderBy, direction] = sortStr?.split(':') ?? '';
-
-  if (
-    (colId === undefined || colId === orderBy) &&
-    (direction === 'asc' || direction === 'desc')
-  ) {
-    return { orderBy, direction } as {
-      orderBy: string;
-      direction: 'asc' | 'desc';
-    };
-  }
-  return {
-    orderBy: undefined,
-    direction: undefined,
-  };
 };
 
 const HeaderCell = ({
