+++ src/app/lib/common.ts
@@ -68,3 +68,7 @@ export const byProp = <
     return aProp > bProp ? (asc ? 1 : -1) : bProp > aProp ? (asc ? -1 : 1) : 0;
   };
 };
+
+export const PAGE_SIZES = [10, 20, 30, 50, 100];
+
+export const DEFAULT_PAGE_SIZE = 30;
